\name{Simulated gene expression data}
\docType{data}
\alias{ss}
\title{Sample variance of log gene expression under two conditions}
\description{
  Simulated gene expression data. 2500 genes under 2 experimental
  conditions, with 8 replicate arrays for each condition. The data is
  presented as mean and sum of squares of the log gene expression, in
  each condition. ss is the matrix containing the sample variances in
  each condition. 
}
\usage{data(ss)}
\format{matrix no. genes x no. experimental conditions}
\keyword{datasets}
