\name{Simulated example data}
\docType{data}
\alias{ybar}
\title{Mean log gene expression under two conditions}
\description{
  Simulated gene expression data. 2500 genes under 2 experimental
  conditions, with 8 replicate arrays for each condition. The data is
  presented as mean and sum of squares of the log gene expression, in
  each condition. ybar is the matrix containing the means in each condition.
}
\usage{data(ybar)}
\format{matrix no. genes x no. experimental conditions}
\keyword{datasets}
